# internal/server/response/auth_responses.go
package response

type Authorized struct {
	AccessToken string `json:"access_token" example:"eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJpZCI6NDIsIm1lc3NhZ2UiOiJJJ20gdGVhcG90In0.S04gkQab6031E4HABjPBBBTt0K5eR1XLn_R4_VI3M3s"`
}